% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/courserafars.R
\docType{package}
\name{courserafars}
\alias{courserafars}
\title{Coursera Fars Package}
\description{
This package is the final assignment of the course.
}
\author{
Mario Morales \email{mario@moralesalfaro.cl}
}
