% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_map_state}
\alias{fars_map_state}
\title{Plot state accidents}
\usage{
fars_map_state(state.num, year)
}
\arguments{
\item{state.num}{A vector indicating the number of the states to plot.}

\item{year}{A vector indicating the years of the accidents to plot.}
}
\value{
This function returns a dataframe. The first column indicates the month, and each
subsequent column shows the number of observations in that month for a specific year.
}
\description{
This function plots the accidents that occured within the selected years in the selected states.
}
\examples{
\dontrun{
fars_map_state(1, 2014)
fars_map_state(5:10, 2013:2015)
}

}
