% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read_years}
\alias{fars_read_years}
\title{Months in data}
\usage{
fars_read_years(years)
}
\arguments{
\item{years}{A vector containing the years to read.}
}
\value{
This function returns a list, in which each element contains the month
for each observation of an specific year.
}
\description{
This function returns the month of each observation for the selected years.
}
\examples{
\dontrun{
fars_read_years(2013:2015)
}

}
