% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_summarize_years}
\alias{fars_summarize_years}
\title{Summarize monthly observations}
\usage{
fars_summarize_years(years)
}
\arguments{
\item{years}{A vector containing the years to read.}
}
\value{
This function returns a dataframe. The first column indicates the month, and each
subsequent column shows the number of observations in that month for a specific year.
}
\description{
This function summarizes the number of observations for each month of the selected years.
}
\examples{
\dontrun{
fars_summarize_years(2013:2015)
}

}
