% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{make_filename}
\alias{make_filename}
\title{Create standard FARS filename}
\usage{
make_filename(year)
}
\arguments{
\item{year}{(YYYY) either character or integer.}
}
\value{
Returns a string: the filename with specified year and format(csv.bz2).
}
\description{
User enters the year (YYYY) to retrieve the standard filename for FARS data.
The output is a string that forms the name of the file as "accident_YYYY.csv.bz".
}
\examples{
\dontrun{make_filename(2020)
make_filename("2020")}

}
