% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_map_state}
\alias{fars_map_state}
\title{Show locations of fatal injuries in a given state and during a given year.}
\usage{
fars_map_state(state.num, year)
}
\arguments{
\item{state.num}{The number of the state to show locations from.}

\item{year}{The year for which to show locations of fatal injuries.}
}
\value{
A map / a plot or an error if there were no fatal injuries to show.
}
\description{
\code{fars_map_state} shows locations of the fatal injuries for a given
state and year on a map.
}
\examples{
\dontrun{
fars_map_state(1, 2015)
}

}
