% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_map_state}
\alias{fars_map_state}
\title{Create a map of accidents in a given state in a given year}
\usage{
fars_map_state(state.num, year)
}
\arguments{
\item{state.num}{An integer corresponding to a state number in the maps package}

\item{year}{A character string or integer indicating a year}
}
\value{
This function produced a plot, if there are accidents to plot in that 
state in that year
}
\description{
This is a function that read an accident files for a given year and
plots a map showing accident locations for a given state in that
year.
}
\note{
If there are no accidents to plot in that state in that year, the function
issues a warning. If the state number does not correspond to a state in the 
accident data file the function produces an error.
}
\examples{
\dontrun{fars_map_state(12,2014)}
}
