% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read}
\alias{fars_read}
\title{Read local FARS data}
\usage{
fars_read(filename)
}
\arguments{
\item{filename}{Path to a FARS csv file (optionally compressed).}
}
\value{
A tibble representing the FARS data or an error
if the file does not exist.
}
\description{
\code{fars_read} reads data from the US National Highway
Traffic Safety Administration's Fatality Analysis Reporting
System (FARS), which is a nationwide census providing the
American public yearly data regarding fatal injuries suffered
in motor vehicle traffic crashes.
}
\examples{
\dontrun{
fars_read("accident_2015.csv.bz2")
fars_read(make_filename(2015))
}

}
