% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read}
\alias{fars_read}
\title{Read a given file into R}
\usage{
fars_read(filename)
}
\arguments{
\item{filename}{A character string giving the name of the file to read into
the session}
}
\value{
This function returns a tbl_df object from the dplyr package
}
\description{
This is a function that reads a given filename, given by the \code{filename}
argument into R using the \code{read_csv} function from the readr 
packages and then returning the read in table as a \code{tbl_df} 
object from the dplyr package.
}
\note{
Calling this function with a filename that does not exist in the working
director will result in an error. This function is called by the make_filename 
function and usually not called on its own.
}
\examples{
\dontrun{fars_read('accident_2013.csv.bz2')}
}
