% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read_years}
\alias{fars_read_years}
\title{Read multiple years of FARS data.}
\usage{
fars_read_years(years)
}
\arguments{
\item{years}{Vector or list of years to read.}
}
\value{
A list of tibbles representing FARS data. If data for a year is
missing, an empty list element is returned together with a warning.
}
\description{
\code{fars_read_years} will read all the given years of FARS data.
}
\examples{
\dontrun{
  fars_read_years(c(2015, 2016))
}

}
