% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read_years}
\alias{fars_read_years}
\title{Read a series of accident files for given years}
\usage{
fars_read_years(years)
}
\arguments{
\item{years}{A vector of years for which to summarize accident months}
}
\value{
This function returns a list of tbl_df object for each year that each
contain the months and year of accidents for that year
}
\description{
This is a function that reads a accident files for a vector of 
a \code{dplyr::tbl_df} with columns for the months and 
given year of accidents in that year.
}
\note{
Calling this function with for a year that doesn't have an accident file
results in an "invalid year" warning for that year. This function is
typically called by the \code{fars_summarize_years()} function and not
run on its own. Hence not exported.
}
\examples{
\\donotrun{fars_read_years(c(2013,2014))}
}
