% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_summarize_years}
\alias{fars_summarize_years}
\title{Number of fatal injuries per month and year.}
\usage{
fars_summarize_years(years)
}
\arguments{
\item{years}{Vector or list of years to read.}
}
\value{
A tibble of monthly accidents per month and year.
}
\description{
\code{fars_summarize_years} summarizes FARS data for the given years
into the number of fatal injuries per month and year.
}
\examples{
\dontrun{
  fars_summarize_years(c(2015, 2016))
}

}
