% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_summarize_years}
\alias{fars_summarize_years}
\title{Summarize counts of a accidents by month for a set of years}
\usage{
fars_summarize_years(years)
}
\arguments{
\item{years}{A vector of years for which to summarize accident months}
}
\value{
This function returns a dplyr::tbl_df object for with a column for
each provided year and a row for each month. The values in the data frame
are counts of accidents for the corresponding month and year.
}
\description{
This is a function that summarizes counts of the number of 
accidents per month for a given set of years from 
accident data files.
}
\examples{
\\donotrun{fars_summarize_years(c(2013,2014)}

}
