% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{make_filename}
\alias{make_filename}
\title{Create a filename for FARS accident data for a given year.}
\usage{
make_filename(year)
}
\arguments{
\item{year}{Year of accident data.}
}
\value{
filename of accident data, e.g. "accident_2015.csv.bz2".
}
\description{
Create a filename for FARS accident data for a given year.
}
\examples{
make_filename(2015)
make_filename("2015")
}
