% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{make_filename}
\alias{make_filename}
\title{Make filename}
\usage{
make_filename(year)
}
\arguments{
\item{year}{Year of the data that the file will contain.}
}
\value{
This function returns a string with the name of the file.
}
\description{
This function returns the name of a filename.
}
\examples{
\dontrun{
make_filename(2017)
}

}
