% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{make_filename}
\alias{make_filename}
\title{Format a valid filename given a year}
\usage{
make_filename(year)
}
\arguments{
\item{year}{A character string or integer giving the year to use to format the filename. This parameter
will be coerced to an integer in the function.}
}
\value{
This function returns a character string of a filename corresponding to the given year
}
\description{
This is a function that reads a given filename, given by the \code{filename}
argument into R using the \code{read_csv} function from the readr 
packages and then returning the read in table as a \code{tbl_df} 
object from the dplyr package.
}
\note{
This function is called by the fars_read_years function and not usually called on its
own. Hence not exported.
}
\examples{
make_filename(2013)
}
